Available Palettes.

This function returns a vector containing the names of all the available palettes in the 'fishualize' package.

fish_palettes()

Value

fish_palettes returns a character vector with the names of the fish palettes available to use.

Examples

fish_palettes()

#>  [1] Acanthurus_leucosternon        Acanthurus_olivaceus          
#>  [3] Acanthurus_sohal               Acanthurus_triostegus         
#>  [5] Antennarius_multiocellatus     Aulostomus_chinensis          
#>  [7] Balistapus_undulatus           Balistoides_conspicillum      
#>  [9] Bodianus_rufus                 Bryaninops_natans             
#> [11] Callanthias_australis          Centropyge_loricula           
#> [13] Cephalopholis_argus            Chaetodon_ephippium           
#> [15] Chaetodon_larvatus             Chaetodon_pelewensis          
#> [17] Chaetodontoplus_conspicillatus Chlorurus_microrhinos         
#> [19] Chromis_vanderbilti            Cirrhilabrus_solorensis       
#> [21] Cirrhilabrus_tonozukai         Clepticus_parrae              
#> [23] Coris_gaimard                  Coryphaena_hippurus           
#> [25] Elagatis_bipinnulata           Epibulus_insidiator           
#> [27] Epinephelus_fasciatus          Epinephelus_striatus          
#> [29] Etheostoma_barrenense          Exallias_brevis               
#> [31] Gomphosus_varius               Gramma_loreto                 
#> [33] Halichoeres_bivittatus         Halichoeres_garnoti           
#> [35] Halichoeres_radiatus           Hemitaurichthys_polylepis     
#> [37] Hexagrammos_lagocephalus       Histiophryne_psychedelica     
#> [39] Holacanthus_ciliaris           Hypoplectrus_puella           
#> [41] Hypsoblennius_invemar          Hypsypops_rubicundus          
#> [43] Lampris_guttatus               Lepomis_megalotis             
#> [45] Lutjanus_sebae                 Melichthys_vidua              
#> [47] Naso_lituratus                 Nemateleotris_magnifica       
#> [49] Odonus_niger                   Oncorhynchus_gorbuscha        
#> [51] Oncorhynchus_keta              Oncorhynchus_kisutch          
#> [53] Oncorhynchus_mykiss            Oncorhynchus_nerka            
#> [55] Oncorhynchus_tshawytscha       Ostorhinchus_angustatus       
#> [57] Ostracion_cubicus              Ostracion_whitleyi            
#> [59] Oxymonacanthus_longirostris    Paracanthurus_hepatus         
#> [61] Parupeneus_insularis           Pomacanthus_imperator         
#> [63] Pomacanthus_xanthometopon      Prionace_glauca               
#> [65] Pseudocheilinus_tetrataenia    Pseudochromis_aldabraensis    
#> [67] Pterois_volitans               Salmo_salar                   
#> [69] Salmo_trutta                   Salvelinus_fontinalis         
#> [71] Scarus_globiceps               Scarus_quoyi                  
#> [73] Sparisoma_viride               Stegastes_nigricans           
#> [75] Synchiropus_splendidus         Taeniura_lymma                
#> [77] Thalassoma_bifasciatum         Thalassoma_hardwicke          
#> [79] Thalassoma_pavo                Trimma_lantana                
#> [81] Valenciennea_strigata          Variola_louti                 
#> [83] Zanclus_cornutus               Zebrasoma_velifer             
#> 84 Levels: Acanthurus_leucosternon Acanthurus_olivaceus ... Zebrasoma_velifer