Available fish silhouettes

This function returns a dataframe containing the all the available fish silhouettes accessible through the 'fishualize' package.

fishapes()

Value

fishapes returns a dataframe containing the all the available fish silhouettes available to use.

Examples

fishapes()

#>                 family                           option
#> 1         Acanthuridae            Acanthurus_nigricauda
#> 2         Acanthuridae            Ctenochaetus_striatus
#> 3         Acanthuridae                   Naso_lituratus
#> 4         Acanthuridae                   Naso_unicornis
#> 5         Acanthuridae                 Zebrasoma_scopas
#> 6         Acanthuridae                Zebrasoma_velifer
#> 7            Alopiidae                 Alopias_vulpinus
#> 8           Apogonidae                 Apogon_binotatus
#> 9         Aulostomidae             Aulostomus_chinensis
#> 10          Balistidae             Balistapus_undulatus
#> 11           Belonidae                    Belone_belone
#> 12          Blenniidae        Antennablennius_adenensis
#> 13          Blenniidae                 Ecsenius_pulcher
#> 14          Blenniidae Hypleurochilus_pseudoaequipinnis
#> 15            Bothidae                    Bothus_mancus
#> 16          Bythitidae    Dinematichthys_iluocoeteoides
#> 17       Callionymidae               Callionymus_bairdi
#> 18          Carangidae                Caranx_melampygus
#> 19      Carcharhinidae                Triaenodon_obesus
#> 20       Centrarchidae                Lepomis_megalotis
#> 21        Chaenopsidae         Acanthemblemaria_spinosa
#> 22      Chaetodontidae                 Chaetodon_auriga
#> 23      Chaetodontidae               Chaetodon_plebeius
#> 24      Chaetodontidae             Chaetodon_vagabundus
#> 25     Channichthyidae          Chaenocephalus_aceratus
#> 26           Cichlidae          Apistogramma_hippolytae
#> 27            Clinidae          Heteroclinus_heptaeolus
#> 28           Clupeidae           Jenkinsia_lamprotaenia
#> 29            Cottidae             Artedius_harringtoni
#> 30     Dactyloscopidae               Gillellus_uranidea
#> 31          Ephippidae                     Platax_teira
#> 32            Esocidae                      Esox_lucius
#> 33          Fundulidae                 Fundulus_grandis
#> 34  Ginglymostomatidae              Nebrius_ferrugineus
#> 35        Gobiesocidae                Acyrtus_lanthanum
#> 36            Gobiidae             Coryogalops_anomolus
#> 37            Gobiidae                   Eviota_guttata
#> 38            Gobiidae                   Eviota_melasma
#> 39            Gobiidae             Fusigobius_neophytus
#> 40            Gobiidae           Gnatholepis_cauerensis
#> 41            Gobiidae                Gobiodon_okinawae
#> 42            Gobiidae           Hetereleotris_vulgaris
#> 43            Gobiidae              Priolepis_hippoliti
#> 44            Gobiidae                      Risor_ruber
#> 45         Grammatidae           Lipogramma_barrettorum
#> 46          Haemulidae          Plectorhinchus_gibbosus
#> 47       Holocentridae             Myripristis_violacea
#> 48       Holocentridae          Sargocentron_spiniferum
#> 49          Kyphosidae             Kyphosus_cinerascens
#> 50            Labridae            Chlorurus_microrhinus
#> 51            Labridae               Chlorurus_sordidus
#> 52            Labridae                    Coris_gaimard
#> 53            Labridae              Epibulus_insidiator
#> 54            Labridae                   Scarus_oviceps
#> 55            Labridae                 Scarus_schlegeli
#> 56        Labrisomidae                Labrisomus_guppyi
#> 57         Lethrinidae              Lethrinus_nebulosus
#> 58           Liparidae            Careproctus_staufferi
#> 59          Lutjanidae                  Lutjanus_gibbus
#> 60           Mobulidae                 Mobula_birostris
#> 61       Monacanthidae                Aluterus_scriptus
#> 62           Mugilidae            Ellochelon_vaigiensis
#> 63            Mullidae        Parupeneus_multifasciatus
#> 64          Muraenidae            Gymnothorax_javanicus
#> 65     Opistognathidae          Opistognathus_aurifrons
#> 66         Ostraciidae     Acanthostracion_quadricornis
#> 67         Ostraciidae                Ostracion_cubicus
#> 68         Pempheridae              Pempheris_oualensis
#> 69            Percidae                 Perca_flavescens
#> 70         Plesiopidae               Assessor_macneilli
#> 71          Poecilidae              Poecilia_reticulata
#> 72       Pomacanthidae              Centropyge_loricula
#> 73       Pomacanthidae            Pomacanthus_imperator
#> 74       Pomacentridae           Abudefduf_sexfasciatus
#> 75       Pomacentridae      Acanthochromis_polyacanthus
#> 76       Pomacentridae               Amphiprion_percula
#> 77       Pomacentridae              Stegastes_nigricans
#> 78       Priacanthidae     Heteropriacanthus_cruentatus
#> 79      Rhincodontidae                  Rhincodon_typus
#> 80          Salmonidae               Oncorhynchus_nerka
#> 81          Scombridae                Thunnus_albacares
#> 82        Scorpaenidae                 Pterois_volitans
#> 83        Scorpaenidae          Scorpaena_jacksoniensis
#> 84        Scorpaenidae                Sebastes_caurinus
#> 85          Serranidae              Cephalopholis_argus
#> 86          Serranidae          Cephalopholis_cruentata
#> 87          Serranidae          Epinephelus_malabaricus
#> 88          Serranidae                Epinephelus_merra
#> 89           Siganidae                Siganus_argenteus
#> 90           Siganidae            Siganus_canaliculatus
#> 91           Siganidae               Siganus_corallinus
#> 92           Siganidae                Siganus_punctatus
#> 93           Siganidae                 Siganus_vulpinus
#> 94        Sphyraenidae              Sphyraena_barracuda
#> 95        Syngnathidae                    Syngnathus_sp
#> 96      Tetraodontidae               Arothron_meleagris
#> 97      Tripterygiidae            Enneanectes_altivelis
#> 98      Tripterygiidae              Enneanectes_matador
#> 99      Tripterygiidae     Enneapterygius_ventermaculus
#> 100     Tripterygiidae           Helcogramma_fuscopinna
#> 101          Zanclidae                 Zanclus_cornutus