Available Palettes.

This function returns a vector containing the names of all the available palettes in the 'fishualize' package.

fish_palettes()

Value

fish_palettes returns a character vector with the names of the fish palettes available to use.

Examples

fish_palettes()

#>   [1] Acanthurus_leucosternon        Acanthurus_olivaceus          
#>   [3] Acanthurus_sohal               Acanthurus_triostegus         
#>   [5] Antennarius_commerson          Antennarius_multiocellatus    
#>   [7] Aulostomus_chinensis           Balistapus_undulatus          
#>   [9] Balistoides_conspicillum       Bodianus_pulchellus           
#>  [11] Bodianus_rufus                 Bryaninops_natans             
#>  [13] Callanthias_australis          Centropyge_loricula           
#>  [15] Cephalopholis_argus            Chaetodon_ephippium           
#>  [17] Chaetodon_larvatus             Chaetodon_pelewensis          
#>  [19] Chaetodontoplus_conspicillatus Chlorurus_microrhinos         
#>  [21] Chlorurus_spilurus             Chromis_vanderbilti           
#>  [23] Cirrhilabrus_solorensis        Cirrhilabrus_tonozukai        
#>  [25] Clepticus_parrae               Coris_gaimard                 
#>  [27] Coryphaena_hippurus            Elagatis_bipinnulata          
#>  [29] Epibulus_insidiator            Epinephelus_fasciatus         
#>  [31] Epinephelus_lanceolatus        Epinephelus_striatus          
#>  [33] Etheostoma_barrenense          Etheostoma_spectabile         
#>  [35] Exallias_brevis                Forcipiger_longirostris       
#>  [37] Gomphosus_varius               Gramma_loreto                 
#>  [39] Halichoeres_bivittatus         Halichoeres_garnoti           
#>  [41] Halichoeres_radiatus           Hemitaurichthys_polylepis     
#>  [43] Hexagrammos_lagocephalus       Histiophryne_psychedelica     
#>  [45] Holacanthus_ciliaris           Hypoplectrus_puella           
#>  [47] Hypsoblennius_invemar          Hypsypops_rubicundus          
#>  [49] Koumansetta_rainfordi          Lampris_guttatus              
#>  [51] Lepomis_megalotis              Lutjanus_sebae                
#>  [53] Melichthys_vidua               Minilabrus_striatus           
#>  [55] Naso_lituratus                 Nemateleotris_magnifica       
#>  [57] Odonus_niger                   Oncorhynchus_gorbuscha        
#>  [59] Oncorhynchus_keta              Oncorhynchus_kisutch          
#>  [61] Oncorhynchus_mykiss            Oncorhynchus_nerka            
#>  [63] Oncorhynchus_tshawytscha       Ostorhinchus_angustatus       
#>  [65] Ostracion_cubicus              Ostracion_whitleyi            
#>  [67] Oxymonacanthus_longirostris    Paracanthurus_hepatus         
#>  [69] Parupeneus_insularis           Pomacanthus_imperator         
#>  [71] Pomacanthus_xanthometopon      Prionace_glauca               
#>  [73] Pseudocheilinus_tetrataenia    Pseudochromis_aldabraensis    
#>  [75] Pterois_volitans               Rhinecanthus_aculeatus        
#>  [77] Rhinecanthus_assasi            Salmo_salar                   
#>  [79] Salmo_trutta                   Salvelinus_fontinalis         
#>  [81] Scarus_ghobban                 Scarus_globiceps              
#>  [83] Scarus_hoefleri                Scarus_quoyi                  
#>  [85] Scarus_tricolor                Serranus_scriba               
#>  [87] Sparisoma_viride               Stegastes_nigricans           
#>  [89] Stegastes_partitus             Stethojulis_bandanensis       
#>  [91] Synchiropus_splendidus         Taeniura_lymma                
#>  [93] Thalassoma_bifasciatum         Thalassoma_hardwicke          
#>  [95] Thalassoma_pavo                Trimma_lantana                
#>  [97] Valenciennea_strigata          Variola_louti                 
#>  [99] Zanclus_cornutus               Zebrasoma_velifer             
#> [101] Zebrasoma_xanthurum           
#> 101 Levels: Acanthurus_leucosternon Acanthurus_olivaceus ... Zebrasoma_xanthurum