Available Palettes.

This function returns a vector containing the names of all the available palettes in the 'fishualize' package.

fish_palettes()

Value

fish_palettes returns a character vector with the names of the fish palettes available to use.

Examples

fish_palettes()

#>  [1] Acanthurus_leucosternon        Acanthurus_olivaceus          
#>  [3] Acanthurus_sohal               Acanthurus_triostegus         
#>  [5] Aulostomus_chinensis           Balistapus_undulatus          
#>  [7] Balistoides_conspicillum       Bryaninops_natans             
#>  [9] Centropyge_loricula            Cephalopholis_argus           
#> [11] Chaetodon_ephippium            Chaetodon_larvatus            
#> [13] Chaetodon_pelewensis           Chaetodontoplus_conspicillatus
#> [15] Chlorurus_microrhinos          Chromis_vanderbilti           
#> [17] Cirrhilabrus_solorensis        Cirrhilabrus_tonozukai        
#> [19] Coris_gaimard                  Coryphaena_hippurus           
#> [21] Elagatis_bipinnulata           Epibulus_insidiator           
#> [23] Epinephelus_fasciatus          Etheostoma_barrenense         
#> [25] Exallias_brevis                Gomphosus_varius              
#> [27] Gramma_loreto                  Halichoeres_radiatus          
#> [29] Hemitaurichthys_polylepis      Hexagrammos_lagocephalus      
#> [31] Histiophryne_psychedelica      Hypsoblennius_invemar         
#> [33] Hypsypops_rubicundus           Lampris_guttatus              
#> [35] Lepomis_megalotis              Lutjanus_sebae                
#> [37] Melichthys_vidua               Naso_lituratus                
#> [39] Nemateleotris_magnifica        Odonus_niger                  
#> [41] Ostorhinchus_angustatus        Ostracion_cubicus             
#> [43] Ostracion_whitleyi             Oxymonacanthus_longirostris   
#> [45] Parupeneus_insularis           Pomacanthus_imperator         
#> [47] Pomacanthus_xanthometopon      Prionace_glauca               
#> [49] Pseudocheilinus_tetrataenia    Pseudochromis_aldabraensis    
#> [51] Pterois_volitans               Salvelinus_fontinalis         
#> [53] Scarus_globiceps               Scarus_quoyi                  
#> [55] Stegastes_nigricans            Synchiropus_splendidus        
#> [57] Taeniura_lymma                 Thalassoma_bifasciatum        
#> [59] Thalassoma_hardwicke           Thalassoma_pavo               
#> [61] Trimma_lantana                 Valenciennea_strigata         
#> [63] Variola_louti                  Zanclus_cornutus              
#> [65] Zebrasoma_velifer             
#> 65 Levels: Acanthurus_leucosternon Acanthurus_olivaceus ... Zebrasoma_velifer